+++ src/main/java/de/tum/in/ase/parser/exception/ParserException.java
@@ -1,4 +1,4 @@
-package de.tum.in.ase.staticCodeAnalysisParser.exception;
+package de.tum.in.ase.parser.exception;
 
 public class ParserException extends Exception {
