+++ src/main/java/de/tum/in/ase/parser/strategy/CheckstyleParser.java
@@ -1,4 +1,4 @@
-package de.tum.in.ase.staticCodeAnalysisParser.strategy;
+package de.tum.in.ase.parser.strategy;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -6,8 +6,8 @@ import java.util.List;
 import nu.xom.Document;
 import nu.xom.Element;
 
-import de.tum.in.ase.staticCodeAnalysisParser.domain.Issue;
-import de.tum.in.ase.staticCodeAnalysisParser.domain.Report;
+import de.tum.in.ase.parser.domain.Issue;
+import de.tum.in.ase.parser.domain.Report;
 
 class CheckstyleParser implements ParserStrategy {
